Backend/requirements.txt
Flask==3.1.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
scikit-learn==1.6.1
pandas==2.3.3
numpy==2.1.1
joblib==1.5.3
requests==2.32.4
gunicorn==23.0.0
PyJWT==2.10.1
